% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{teste_t_pareado_interativo}
\alias{teste_t_pareado_interativo}
\title{Teste t Pareado Interativo}
\usage{
teste_t_pareado_interativo()
}
\value{
Um objeto de classe "htest" com os resultados do teste t pareado
}
\description{
Esta funcao realiza um teste t pareado de forma interativa, solicitando
os dados das duas variaveis e o nivel de significancia desejado.
}
\examples{
\dontrun{
teste_t_pareado_interativo()
}
}
