% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\docType{package}
\name{ComparMedia}
\alias{ComparMedia-package}
\alias{ComparMedia}
\title{ComparMedia: Teste t Pareado Interativo}
\description{
Este pacote fornece uma funcao interativa para realizar testes t pareados,
facilitando a analise estatistica de dados pareados atraves de uma interface
amigavel no console R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WpereiraPA/Comparar2Medias}
  \item Report bugs at \url{https://github.com/WpereiraPA/Comparar2Medias/issues}
}

}
\author{
\strong{Maintainer}: Wanderley Xavier Pereira \email{wander.wx@gmail.com}

}
